feat: second time payment flow
