+++ src/components/organisms/CreditCardForm.tsx
@@ -1,65 +1,40 @@
 import { ReactElement, useState, useEffect } from 'react';
-import { Container } from '@/styles/components/organisms/CreditCardForm';
-import Input from '@/components/atom/Input';
-import cvvIcon from '@/assets/components/organisms/CreditCardForm/cvv.svg';
+import {
+  Container,
+  MainCard,
+  Card,
+  ModalContainer,
+  CardContainer,
+  CloseDiv,
+} from '@/styles/components/organisms/CreditCardForm';
 import Dropdown from '@/components/atom/Dropdown';
-import diners from '@/assets/components/atoms/Input/diners-club.svg';
-import amex from '@/assets/components/atoms/Input/american-express.svg';
 import visa from '@/assets/components/atoms/Input/visa.svg';
 import discover from '@/assets/components/atoms/Input/discover.svg';
-import generic from '@/assets/components/atoms/Input/generic.svg';
 import mastercard from '@/assets/components/atoms/Input/mastercard.svg';
-import BinSearch from '@/@types/BinSearch';
 import { usePayment } from '@/hooks/payment';
 import { useBag } from '@/hooks/bag';
 import formatValueWo$ from '@/utils/formatValueWo$';
-import axios, { AxiosResponse } from 'axios';
 import pagarme from 'pagarme';
 import Api from '@/services/api';
 import { useAuth } from '@/hooks/auth';
 import { useRouter } from 'next/router';
+import Modal from '@/components/organisms/Modal';
+import { FiCheck } from 'react-icons/fi';
+import CreditCardFields from '@/components/molecule/CreditCardFields';
+import { MdClose } from 'react-icons/md';
 import Button from '../atom/Button';
 
-const creditCardBrands = {
-  amex,
-  visa,
-  discover,
-  generic,
-  mastercard,
-  diners,
-};
-
-const months = [
-  { id: 1, label: '01', value: '01' },
-  { id: 2, label: '02', value: '02' },
-  { id: 3, label: '03', value: '03' },
-  { id: 4, label: '04', value: '04' },
-  { id: 5, label: '05', value: '05' },
-  { id: 6, label: '06', value: '06' },
-  { id: 7, label: '07', value: '07' },
-  { id: 8, label: '08', value: '08' },
-  { id: 9, label: '09', value: '09' },
-  { id: 10, label: '10', value: '10' },
-  { id: 11, label: '11', value: '11' },
-  { id: 12, label: '12', value: '12' },
-];
-
-const currentYear = new Date().getFullYear() - 2000;
-
-const years = Array.from(Array(50), (_, index) => ({
-  id: index,
-  value: `20${index + currentYear}`,
-  label: `20${index + currentYear}`,
-}));
-
 const CreditCardForm = (): ReactElement => {
-  const [creditCardBrand, setCreditCardBrand] = useState('generic');
   const [isPaymentButtonDisabled, setIsPaymentButtonDisabled] = useState(true);
+  const [displayChangeCreditCard, setDisplayChangeCreditCard] = useState(false);
+  const [displayAddNewCard, setDisplayAddNewCard] = useState(false);
   const { paymentData, setPaymentData } = usePayment();
   const { bagTotalPrice, bagItems } = useBag();
   const { user } = useAuth();
 
   const router = useRouter();
+
+  const alreadyHasCreditCard = true;
 
   useEffect(() => {
     if (
@@ -88,30 +63,6 @@ const CreditCardForm = (): ReactElement => {
     value: `${index + 1}x de R$${formatValueWo$(bagTotalPrice / (index + 1))}`,
     label: `${index + 1}x de R$${formatValueWo$(bagTotalPrice / (index + 1))}`,
   }));
-
-  function handleCreditCardNumberOnChange(value): void {
-    const creditCardNumber = value.replace(/[\s*/_*/]/gm, '');
-
-    setPaymentData({
-      ...paymentData,
-      card: { ...paymentData.card, card_number: creditCardNumber },
-    });
-
-    if (creditCardNumber.length === 6 && creditCardNumber !== '') {
-      axios
-        .get(`https://lookup.binlist.net/${creditCardNumber}`, {
-          headers: {
-            'Accept-Version': '3',
-          },
-        })
-        .then((res: AxiosResponse<BinSearch>) =>
-          setCreditCardBrand(res.data.scheme),
-        )
-        .catch(() => setCreditCardBrand('generic'));
-    } else if (creditCardNumber.length < 6) {
-      setCreditCardBrand('generic');
-    }
-  }
 
   function handlePaymentClick(): void {
     const items = [];
@@ -196,84 +147,87 @@ const CreditCardForm = (): ReactElement => {
 
   return (
     <Container className="credit-card-form">
-      <Input
-        className="input-credit"
-        type="text"
-        label="Número do cartão"
-        value={paymentData.card?.card_number}
-        name="ccnum"
-        mask="9999 9999 9999 9999"
-        onChange={event => handleCreditCardNumberOnChange(event.target.value)}
-        x-autocompletetype="cc-number"
-        iconAfter={creditCardBrands[creditCardBrand]}
-      />
-      <Input
-        className="input-credit"
-        type="text"
-        name="ccname"
-        value={paymentData.card?.card_holder_name}
-        onChange={event =>
-          setPaymentData({
-            ...paymentData,
-            card: { ...paymentData.card, card_holder_name: event.target.value },
-          })
-        }
-        label="Nome impresso no cartão"
-      />
-      <div className="card-expiration-div">
-        <div>
-          <Dropdown
-            defaultValue="Mês"
-            type="small"
-            options={months}
-            setValue={value =>
-              setPaymentData({
-                ...paymentData,
-                card: { ...paymentData.card, card_expiration_month: value },
-              })
-            }
-            value={paymentData.card?.card_expiration_month}
-            className="small first-dropdown"
-          />
-        </div>
-        <div>
-          <Dropdown
-            defaultValue="Ano"
-            type="small"
-            options={years}
-            setValue={value =>
-              setPaymentData({
-                ...paymentData,
-                card: { ...paymentData.card, card_expiration_year: value },
-              })
-            }
-            value={paymentData.card?.card_expiration_year}
-            className="small first-dropdown"
-          />
-        </div>
-      </div>
-      <div className="cvv-div">
-        <div>
-          <Input
-            className="input-credit"
-            type="text"
-            name="cvv"
-            value={paymentData.card?.card_cvv}
-            onChange={event => {
-              const cvv = event.target.value.replace(/[\s*/_*/]/gm, '');
-              setPaymentData({
-                ...paymentData,
-                card: { ...paymentData.card, card_cvv: cvv },
-              });
-            }}
-            label="CVV"
-            mask="999"
-          />
-        </div>
-        <div>
-          <img src={cvvIcon} alt="ícone de cvv" />
-        </div>
-      </div>
+      {alreadyHasCreditCard ? (
+        <>
+          <MainCard>
+            <img src={mastercard} alt="Ícone da marca do cartão de crédito" />
+            <div>
+              <span>MASTERCARD *1234</span>
+              <caption>Bruno Gomes da Cruz</caption>
+            </div>
+            <button
+              type="button"
+              onClick={() => setDisplayChangeCreditCard(true)}
+            >
+              Trocar
+            </button>
+          </MainCard>
+          <Modal
+            isOpen={displayChangeCreditCard}
+            setIsOpen={() => setDisplayChangeCreditCard(false)}
+          >
+            <CloseDiv>
+              <MdClose onClick={() => setDisplayChangeCreditCard(false)} />
+            </CloseDiv>
+            <ModalContainer>
+              {displayAddNewCard ? (
+                <>
+                  <h3>Adicionar cartão de crédito</h3>
+                  <div className="container-credit-card-fields">
+                    <CreditCardFields isModal />
+                    <Button>Adicionar cartão de crédito</Button>
+                  </div>
+                </>
+              ) : (
+                <>
+                  <h3>Com qual cartão você prefere pagar?</h3>
+                  <CardContainer>
+                    <Card>
+                      <img
+                        src={mastercard}
+                        alt="Ícone da marca do cartão de crédito"
+                      />
+                      <div>
+                        <span>9876 54XX XXXX 1234</span>
+                      </div>
+                      <FiCheck className="check-icon" />
+                    </Card>
+                    <Card>
+                      <img
+                        src={discover}
+                        alt="Ícone da marca do cartão de crédito"
+                      />
+                      <div>
+                        <span>9876 54XX XXXX 1234</span>
+                      </div>
+                    </Card>
+                    <Card>
+                      <img
+                        src={visa}
+                        alt="Ícone da marca do cartão de crédito"
+                      />
+                      <div>
+                        <span>9876 54XX XXXX 1234</span>
+                      </div>
+                    </Card>
+                  </CardContainer>
+                  <button
+                    type="button"
+                    className="add-new-credit-card"
+                    onClick={() => setDisplayAddNewCard(true)}
+                  >
+                    Adicionar cartão de crédito
+                  </button>
+
+                  <Button>Selecionar cartão de crédito</Button>
+                </>
+              )}
+            </ModalContainer>
+          </Modal>
+        </>
+      ) : (
+        <CreditCardFields />
+      )}
       <div className="installments-container">
         <Dropdown
           defaultValue="Parcelar em"
